requirements.txt
antiorm==1.2.1
certifi==2020.12.5
cffi==1.14.4
cryptography==3.2.1
db==0.1.1
db-sqlite3==0.0.1
dnspython==2.0.0
future==0.18.2
pycparser==2.20
pymongo==3.11.2
python-telegram-bot==12.3.0
six==1.15.0
tornado==6.1
